pyproject.toml
[project]
name = "who-am-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "channels[redis]>=4.2.2",
    "channels-redis>=4.2.1",
    "daphne>=4.2.1",
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "google-generativeai>=0.5.0",
    "langchain>=0.2.0",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "djangorestframework-simplejwt>=5.5.0",
]
